fix(0.19): expiryTimestamp lock-basis + operatorCount width + revision-read tests
Addresses the blocking review findings on the 0.19 adoption:

- CRITICAL: withdrawal lock displays were reading LockInfo.expiryBlock, which the synced
  0.19 ABI renamed to expiryTimestamp — viem keys decoded tuples by ABI component name, so
  every read was undefined and all locks showed as unlocked. Read expiryTimestamp on every
  revision; branch the comparison basis: legacy stores a BLOCK NUMBER (vs currentBlockNumber),
  0.18/0.19 store a unix TIMESTAMP (vs wall-clock seconds). Same uint64 slot, only the value
  semantics differ.
- HIGH: legacy getBlueprint ABI had operatorCount as uint64; the real pre-#194 struct is
  uint32. Wrong width mis-aligned the following membership/pricing/active fields on legacy.
- MED: blueprintSources -> blueprintSourcesHash — grep confirms zero dapp callers of either;
  documented the rename (the setBlueprintSources WRITE is unchanged).
- MED: canScheduleExit view removed -> fail-closed on v019; added TODO(v019) markers at the
  fallback and the exit-button block to track the getExitStatus follow-up.
- MED: extracted the pure decode/ABI-select logic out of the wagmi-importing hooks into
  wagmi-free modules (slashProposal.ts, binaryVersion.ts, re-exported, zero consumer change)
  and added 11 revision tests: v019 nulls binaryUri/reportUri without throwing; v018 decodes
  the full 14-field slash tuple; v019 short-tuple positional decode has no index slip.
- MED: documented the v019 LockInfo/expiryTimestamp semantics in getTntCoreRevisionByChainId.

Verified: nx typecheck (tangle-dapp + tangle-cloud) pass; nx test tangle-shared-ui 85 pass
(11 new + 74 existing); prettier clean; no .expiryBlock runtime reads remain.

+++ libs/tangle-shared-ui/src/data/blueprints/readBlueprintCore.ts
@@ -25,7 +25,11 @@ const GET_BLUEPRINT_LEGACY_ABI = [
           { name: 'owner', type: 'address' },
           { name: 'manager', type: 'address' },
           { name: 'createdAt', type: 'uint64' },
-          { name: 'operatorCount', type: 'uint64' },
+          // Legacy `Types.Blueprint.operatorCount` is `uint32`, not `uint64`
+          // (confirmed against the pre-#194 Solidity struct). A wrong width
+          // here mis-aligns the packed tail (membership/pricing/active), so it
+          // must match the on-chain type exactly.
+          { name: 'operatorCount', type: 'uint32' },
           { name: 'membership', type: 'uint8' },
           { name: 'pricing', type: 'uint8' },
           { name: 'active', type: 'bool' },
